add lib/processing/capture.rb
