+++ lib/processing.rb
@@ -14,4 +14,5 @@ require 'processing/image'
 require 'processing/font'
 require 'processing/touch'
 require 'processing/shader'
+require 'processing/capture'
 require 'processing/processing'
